rag_tools/requirements.txt
aiohttp>=3.8.5
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dateutil>=2.8.2
python-dotenv>=1.0.0
rich>=13.5.2
supabase>=1.0.3
tenacity>=8.2.3
langchain>=0.0.267
langchain-community>=0.0.10
langchain-groq==0.2.4
langchain-mistralai>=0.0.1
langchain-text-splitters>=0.0.1 